+++ FreshdeskApi.Client/Companies/FreshdeskCompaniesClient.cs
@@ -10,6 +10,7 @@ using FreshdeskApi.Client.Companies.Models;
 using FreshdeskApi.Client.Companies.Requests;
 using FreshdeskApi.Client.Extensions;
 using FreshdeskApi.Client.Models;
+using FreshdeskApi.Client.Pagination;
 
 namespace FreshdeskApi.Client.Companies;
