+++ FreshdeskApi.Client/Products/FreshdeskProductClient.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using FreshdeskApi.Client.Extensions;
 using FreshdeskApi.Client.Models;
+using FreshdeskApi.Client.Pagination;
 using FreshdeskApi.Client.Products.Models;
 
 namespace FreshdeskApi.Client.Products;
