+++ FreshdeskApi.Client/Tickets/FreshdeskTicketClient.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using FreshdeskApi.Client.Conversations.Models;
 using FreshdeskApi.Client.Extensions;
 using FreshdeskApi.Client.Models;
+using FreshdeskApi.Client.Pagination;
 using FreshdeskApi.Client.Tickets.Models;
 using FreshdeskApi.Client.Tickets.Requests;
